Runtime refactoring
