+++ tyrian/js/src/main/scala/tyrian/runtime/ModelHolder.scala
@@ -1,0 +1,3 @@
+package tyrian.runtime
+
+final class ModelHolder[Model](val model: Model, val updated: Boolean)
