+++ tyrian/js/src/main/scala/tyrian/runtime/TyrianRuntime.scala
@@ -8,31 +8,21 @@ import cats.syntax.all.*
 import fs2.Stream
 import org.scalajs.dom
 import org.scalajs.dom.Element
-import snabbdom._
-import snabbdom.modules._
+import snabbdom.VNode
 import tyrian.Attr
 import tyrian.Attribute
 import tyrian.Cmd
 import tyrian.Event
 import tyrian.Html
-import tyrian.NamedAttribute
-import tyrian.Property
 import tyrian.Sub
-import tyrian.Tag
-import tyrian.Text
-import util.Functions.fun
-
-import scala.scalajs.js
-import scala.scalajs.js.Dynamic.{literal => obj}
 
 final class TyrianRuntime[F[_]: Async, Model, Msg](
     initCmd: Cmd[F, Msg],
     update: Model => Msg => (Model, Cmd[F, Msg]),
     view: Model => Html[Msg],
     subscriptions: Model => Sub[F, Msg],
-    node: Element,
     model: Ref[F, ModelHolder[Model]],
-    vnode: Ref[F, Option[VNode]],
+    vnode: Ref[F, Element | VNode],
     queue: => Queue[F, F[Unit]],
     dispatcher: => Dispatcher[F]
 ):
@@ -53,23 +43,17 @@ final class TyrianRuntime[F[_]: Async, Model, Msg](
       for {
         currentModel <- model.get
         (updatedState, cmd) = update(currentModel.model)(msg)
-        complete <- completeUpdate(cmd, updatedState)
-      } yield complete
+        _ <- completeUpdate(cmd, updatedState)
+      } yield ()
 
     dispatcher.unsafeRunAndForget(res)
 
   private def completeUpdate(cmd: Cmd[F, Msg], updatedState: Model): F[Unit] =
-    val results: F[Stream[F, F[Unit]]] =
-      for {
-        _           <- model.set(ModelHolder(updatedState, true))
-        sideEffects <- gatherSideEffects(cmd, subscriptions(updatedState))
-      } yield Stream.emits(sideEffects)
-
-    Async[F].flatMap(results) { (stream: Stream[F, F[Unit]]) =>
-      stream.foreach(queue.offer).compile.drain
-    }
-
-  private given CanEqual[Option[_], Option[_]] = CanEqual.derived
+    for {
+      _           <- model.set(ModelHolder(updatedState, true))
+      sideEffects <- gatherSideEffects(cmd, subscriptions(updatedState))
+      _           <- Stream.emits(sideEffects).foreach(queue.offer).compile.drain
+    } yield ()
 
   private def gatherSideEffects(
       cmd: Cmd[F, Msg],
@@ -123,83 +107,28 @@ final class TyrianRuntime[F[_]: Async, Model, Msg](
 
   // Render
 
-  private def toVNode(html: Html[Msg]): VNode =
-    html match
-      case Tag(name, attrs, children) =>
-        val as: List[(String, String)] =
-          attrs.collect {
-            case Attribute(n, v)   => (n, v)
-            case NamedAttribute(n) => (n, "")
-          }
-
-        val props: List[(String, PropValue)] =
-          attrs.collect { case Property(n, v) => (n, v) }
-
-        val events: List[(String, EventHandler)] =
-          attrs.collect { case Event(n, msg) =>
-            (n, EventHandler((e: dom.Event) => onMsg(msg.asInstanceOf[dom.Event => Msg](e))))
-          }
-
-        val data: VNodeData =
-          VNodeData.empty.copy(
-            props = props.toMap,
-            attrs = as.toMap,
-            on = events.toMap
-          )
-
-        val childrenElem: Array[VNode] =
-          children.toArray.map {
-            case t: Text            => VNode.text(t.value)
-            case subHtml: Html[Msg] => toVNode(subHtml)
-          }
-
-        h(name, data, childrenElem)
-
-  private lazy val patch: Patch =
-    snabbdom.init(
-      Seq(
-        Attributes.module,
-        Classes.module,
-        Props.module,
-        Styles.module,
-        EventListeners.module,
-        Dataset.module
-      )
-    )
-
-  private def render(oldNode: Element | VNode, model: Model): VNode =
-    oldNode match
-      case em: Element => patch(em, toVNode(view(model)))
-      case vn: VNode   => patch(vn, toVNode(view(model)))
-
-  private def renderLoop(time: Double): Unit =
-    def res: F[Unit] =
+  private def renderLoop(): Unit =
+    def redraw: F[Unit] =
       Async[F].flatMap(model.get) { m =>
         if m.updated then
           for {
-            n <- vnode.get
-            _ <- vnode.set(n match {
-              case Some(existingNode) => Some(render(existingNode, m.model))
-              case None               => Some(render(node, m.model))
-            })
+            _ <- vnode.updateAndGet(n => Rendering.render(n, m.model, view, onMsg))
             _ <- model.set(ModelHolder(m.model, false))
           } yield ()
         else Async[F].unit
       }
 
-    dom.window.requestAnimationFrame { t =>
-      dispatcher.unsafeRunAndForget(res)
-      renderLoop(t)
+    dom.window.requestAnimationFrame { _ =>
+      dispatcher.unsafeRunAndForget(redraw)
+      renderLoop()
     }
 
   // Start up
 
   def start(): Unit =
-    renderLoop(0)
+    renderLoop()
     dispatcher.unsafeRunAndForget(
       model.get.flatMap { m =>
         completeUpdate(initCmd, m.model)
       }
     )
-
-final class ModelHolder[Model](val model: Model, val updated: Boolean)
